puts all styles and components in same file"
