--- src/Components/Menu/styles.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { styled } from '@material-ui/styles';
-import {IconButton} from '@material-ui/core';
-
-
-export const StyledIconButton = styled(IconButton)({
-    position: 'absolute',
-    right: 0,
-    color: 'white',
-    'z-index': 999
-});
