+++ src/Components/SideBar.js
@@ -1,7 +1,13 @@
 import React, {useState} from 'react'
 import {Slide} from "@material-ui/core"
-import {StyledAside} from "./styles"
+import styled from "styled-components"
 
+const StyledAside = styled.aside`
+   background: black;
+   border-left: 2px solid white;
+   grid-column: 10 / span 3;
+   grid-row: 1 / span 12;
+`
 const Sidebar = () => {
     const [visible, setVisible] = useState(true)
     return (
