--- src/Components/Sidebar/styles.js
@@ -1,7 +1,0 @@
-import styled from "styled-components"
-
-export const StyledAside = styled.aside`
-   background: black;
-   border-left: 2px solid white;
-   grid-column: 7 / span 2
-`
