--- src/Components/World/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import {GameArea} from "./styles"
-const World = () => {
-    return ( 
-        <GameArea />
-     );
-}
- 
-export default World;
