--- src/Components/World/styles.js
@@ -1,5 +1,0 @@
-import styled from "styled-components"
-
-export const GameArea = styled.div`
-    background: white
-`
