--- src/Components/WorldPage/styles.js
@@ -1,9 +1,0 @@
-import styled from "styled-components"
-
-export const StyledMain = styled.main`
-    margin: 0;
-    min-height: 100vh;
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr;
-    background: black;
-`
