--- src/Components/index.js
@@ -1,1 +1,0 @@
-export {default as WorldPage} from './WorldPage';
